algunos cambios

+++ index.js
@@ -11,8 +11,35 @@ input.addEventListener("input", function () {
 })
 
 botonGuardar.addEventListener("click", () => {
-    alert("Nombre guardado: " + nombre)
-    localStorage.setItem("nombre",nombre)
+    if (localStorage.getItem("nombre")) {
+        Swal.fire({
+            title: "Ya existe un nombre guardado",
+            text: "Desea sobrescribirlo?",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            cancelButtonText: "No",
+            confirmButtonText: "Sí"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                localStorage.setItem("nombre", nombre)
+                Swal.fire({
+                    title: "Guardado!",
+                    text: "El nombre de usuario ha sido guardado.",
+                    icon: "success"
+                });
+            }
+        })
+    }
+    else {
+        localStorage.setItem("nombre", nombre)
+        Swal.fire({
+            title: "Guardado!",
+            text: "El nombre de usuario ha sido guardado.",
+            icon: "success"
+        });
+    }
 })
 
 boton.addEventListener("click",() => {
